third_party/blink/web_tests/external/wpt/editing/run/indent-expected.txt

This is a testharness.js-based test.
Found 71 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["indent",""]] "<div><p>Foo[barbaz<p dir=rtl>פובר]בז</div><p>Extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><div><p>Foobarbaz</p></div><div><p dir=\\"rtl\\">פוברבז</p></div></blockquote><p>Extra</p>" not in array ["<blockquote><div><p>Foobarbaz</p><p dir=\\"rtl\\">פוברבז</p></div></blockquote><p>Extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><div><p>Foobarbaz</p><p dir=\\"rtl\\">פוברבז</p></div></blockquote><p>Extra</p>"]
[FAIL] [["indent",""]] "{<p><p> <p>foo</p>}<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p></p><p> </p><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p></blockquote><p>extra</p>" not in array ["<blockquote><p></p><p> </p><p>foo</p></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p></p><p> </p><p>foo</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<p>{}<br>foo</p><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p><br></p></blockquote>foo</p><p>extra</p>" not in array ["<blockquote><br></blockquote><p>foo</p><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><br></blockquote><p>foo</p><p>extra</p>"]
[FAIL] [["indent",""]] "<span>{}<br>foo</span>bar<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<span><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><span><br></span></blockquote>foo</span>bar<p>extra</p>" not in array ["<blockquote><br></blockquote><span>foo</span>bar<p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><br></blockquote><span>foo</span>bar<p>extra</p>"]
[FAIL] [["indent",""]] "<span>foo<br>{}</span>bar<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<span>foo<br></span><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote><p>extra</p>" not in array ["<span>foo</span><blockquote>bar</blockquote><p>extra</p>", "<span>foo</span><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<p>foo</p>{}<p>bar</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "<p>foo</p><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote>"
[FAIL] [["indent",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<table><tbody><tr><td><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo</blockquote></td><td><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote></td><td>baz</td></tr></tbody></table><p>extra</p>" not in array ["<table><tbody><tr><td><blockquote>foo</blockquote></td><td><blockquote>bar</blockquote></td><td>baz</td></tr></tbody></table><p>extra</p>", "<table><tbody><tr><td><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo</blockquote></td><td><blockquote>bar</blockquote></td><td>baz</td></tr></tbody></table><p>extra</p>"]
[FAIL] [["indent",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<table><tbody><tr><td><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo</blockquote></td><td><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote></td><td><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">baz</blockquote></td></tr></tbody></table><p>extra</p>" not in array ["<blockquote><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><table><tbody><tr><td>foo</td><td>bar</td><td>baz</td></tr></tbody></table></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "foobar<br>[ba]z<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "foobar<br><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">baz</blockquote><p>extra</p>" not in array ["foobar<blockquote>baz</blockquote><p>extra</p>", "foobar<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">baz</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<div>foo<p>[bar]</p>baz</div><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div>foo</div><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><div><p>bar</p></div></blockquote><div>baz</div><p>extra</p>" not in array ["<div>foo<blockquote><p>bar</p></blockquote>baz</div><p>extra</p>", "<div>foo<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"<p>bar</p></blockquote>baz</div><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote><p>foobar</p></blockquote></blockquote><blockquote><p>baz</p></blockquote><p>extra</p>" not in array ["<blockquote><blockquote><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>", "<blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote><p>foobar</p></blockquote><blockquote><p>baz</p></blockquote></blockquote><p>extra</p>" not in array ["<blockquote><blockquote><p>foobar</p><p>baz</p></blockquote></blockquote><p>extra</p>", "<blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p><p>baz</p></blockquote></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote><p>foo[bar]</p></blockquote><p>baz</p><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote><p>foobar</p></blockquote></blockquote><p>baz</p><p>extra</p>" not in array ["<blockquote><blockquote><p>foobar</p></blockquote></blockquote><p>baz</p><p>extra</p>", "<blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote></blockquote><p>baz</p><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote><p>foo[bar</p></blockquote><p>b]az</p><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>" not in array ["<blockquote><blockquote><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>", "<blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p></blockquote><blockquote><p>bar</p></blockquote><p>extra</p>" not in array ["<blockquote><p>foo</p><p>bar</p></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p><p>bar</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<p>foo<blockquote><p>bar</blockquote><p>[baz]<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p>foo</p><blockquote><p>bar</p></blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>" not in array ["<p>foo</p><blockquote><p>bar</p><p>baz</p></blockquote><p>extra</p>", "<p>foo</p><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p><p>baz</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote><p>foo</blockquote><p>[bar]<blockquote><p>baz</blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote><p>foo</p></blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote><blockquote><p>baz</p></blockquote><p>extra</p>" not in array ["<blockquote><p>foo</p><p>bar</p><p>baz</p></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p><p>bar</p><p>baz</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote>foo[bar]<br>baz</blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote>foobar</blockquote></blockquote><blockquote>baz</blockquote><p>extra</p>" not in array ["<blockquote><blockquote>foobar</blockquote>baz</blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote>foobar</blockquote>baz</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote>foo[bar<br>b]az</blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote>foobar</blockquote><blockquote>baz</blockquote></blockquote><p>extra</p>" not in array ["<blockquote><blockquote>foobar<br>baz</blockquote></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote>foobar<br>baz</blockquote></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "[foo]<blockquote>bar</blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo</blockquote><blockquote>bar</blockquote><p>extra</p>" not in array ["<blockquote>foo<br>bar</blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo<br>bar</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "foo<blockquote>bar</blockquote>[baz]<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "foo<blockquote>bar</blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">baz</blockquote><p>extra</p>" not in array ["foo<blockquote>bar<br>baz</blockquote><p>extra</p>", "foo<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar<br>baz</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "[foo<blockquote>bar</blockquote>baz]<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo<blockquote>bar</blockquote>baz</blockquote><p>extra</p>" not in array ["<blockquote>foo<blockquote>bar</blockquote>baz</blockquote><p>extra</p>", "<blockquote>foo<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote>baz</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote>foo</blockquote>[bar]<blockquote>baz</blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote>foo</blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote><blockquote>baz</blockquote><p>extra</p>" not in array ["<blockquote>foo<br>bar<br>baz</blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo<br>bar<br>baz</blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foobar</p></blockquote></blockquote><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>baz</p></blockquote><p>extra</p>" not in array ["<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><blockquote><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>", "<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foobar</p></blockquote><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>baz</p></blockquote></blockquote><p>extra</p>" not in array ["<blockquote><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foobar</p><p>baz</p></blockquote></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foobar</p><p>baz</p></blockquote></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<p>[foo]<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>bar</blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foo</p><p>bar</p></blockquote><p>extra</p>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p></blockquote><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>bar</p></blockquote><p>extra</p>"
[FAIL] [["indent",""]] "<p>foo<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>bar</blockquote><p>[baz]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>bar</p><p>baz</p></blockquote><p>extra</p>" but got "<p>foo</p><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>bar</p></blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["indent",""]] "<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>foo</blockquote><p>[bar]<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>baz</blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foo</p><p>bar</p><p>baz</p></blockquote><p>extra</p>" but got "<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foo</p></blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["indent",""]] "<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>foo</blockquote><p>[bar]<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>baz</blockquote><p>extra" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["indent",""]] "<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>foo</blockquote><p>[bar]<blockquote style=\\"margin-right: 0\\" dir=\\"ltr\\"><p>baz</blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foo</p><p>bar</p><p>baz</p></blockquote><p>extra</p>" but got "<blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>foo</p></blockquote><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote><blockquote style=\\"margin-right:0\\" dir=\\"ltr\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["indent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote></blockquote><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>" not in array ["<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>", "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote><p>baz</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p></blockquote><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote></blockquote><p>extra</p>" not in array ["<blockquote><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p><p>baz</p></blockquote></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foobar</p><p>baz</p></blockquote></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote>f[oo<blockquote>b]ar</blockquote></blockquote><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote>foo</blockquote><blockquote><blockquote>bar</blockquote></blockquote></blockquote><p>extra</p>" not in array ["<blockquote><blockquote>foo<blockquote>bar</blockquote></blockquote></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><blockquote>foo<blockquote>bar</blockquote></blockquote></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<ol><li>foo</ol>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><blockquote>bar</blockquote>" but got "<ol><li>foo</li></ol><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">bar</blockquote>"
[FAIL] [["indent",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo<br>bar</li></ol><li>baz</li></ol>" but got "<ol><ol><li>bar</li></ol><li>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li><div>foo</div>bar</li></ol><li>baz</li></ol>" but got "<ol><li><div><blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foo</blockquote></div>bar</li><li>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>[bar]<li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><ol><li>bar</li></ol><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo<ol><ol><li>bar</li></ol><li>baz</li></ol></li><li>quz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>bar<li>[baz]</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li><ol><li>baz</li></ol></ol><li>quz</li></ol>" but got "<ol><li>foo<ol><li>bar</li><ol><li>baz</li></ol></ol></li><li>quz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>b[a]r</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><ol><li>bar</li></ol></ol><li>baz</li></ol>" but got "<ol><li>foo<ol><ol><li>bar</li></ol></ol></li><li>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo{<ol><li>bar</ol>}<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><ol><li>bar</li></ol></ol><li>baz</li></ol>" but got "<ol><ol><li>foo<ol><li>bar</li></ol></li></ol><li>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>[foo]<ol><li>bar</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li><li>bar</li></ol><li>baz</li></ol>" but got "<ol><ol><li>foo<ol><li>bar</li></ol></li></ol><li>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<li>[bar]<ol><li>baz</ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol><li>quz</li></ol>" but got "<ol><li>foo</li><ol><li>bar<ol><li>baz</li></ol></li></ol><li>quz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>bar<li>baz</ol><li>[quz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li><li>baz</li><li>quz</li></ol></ol>" but got "<ol><li>foo<ol><li>bar</li><li>baz</li></ol></li><ol><li>quz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><ol id=u1><li id=i1>foo</ol><li id=i2>[bar]</li><ol id=u3><li id=i3>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol id=\\"u1\\"><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li><li id=\\"i3\\">baz</li></ol></ol>" but got "<ol><ol id=\\"u3\\"><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li><li id=\\"i3\\">baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><ol><li id=i1>foo</ol><li id=i2>[bar]</li><ol id=u3><li id=i3>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li><li id=\\"i3\\">baz</li></ol></ol>" but got "<ol><ol id=\\"u3\\"><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li><li id=\\"i3\\">baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><ol id=u1><li id=i1>foo</ol><li id=i2>[bar]</li><ol><li id=i3>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol id=\\"u1\\"><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li><li id=\\"i3\\">baz</li></ol></ol>" but got "<ol><ol><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li><li id=\\"i3\\">baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><ol id=u1><li id=i1>foo</ol><li id=i2>[bar]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol id=\\"u1\\"><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li></ol></ol>" but got "<ol><ol><li id=\\"i1\\">foo</li><li id=\\"i2\\">bar</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>[foo<ol><li>bar]</ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li><ol><li>bar</li></ol></ol><li>baz</li></ol>" but got "<ol><ol><li>foo<ol><li>bar</li></ol></li></ol><li>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>b[ar</ol><li>b]az</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><ol><li>bar</li></ol><li>baz</li></ol></ol>" but got "<ol><li>foo<ol><ol><li>bar</li></ol></ol></li><ol><li>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>[foo<ol><li>bar</ol><li>baz]</ol><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<ol><ol><li>foo<ol><li>bar</li></ol></li><li>baz</li></ol></ol><p>extra</p>" not in array ["<blockquote><ol><li>foo</li><ol><li>bar</li></ol><li>baz</li></ol></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><ol><li>foo</li><ol><li>bar</li></ol><li>baz</li></ol></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<ol><li>[foo</li><ol><li>bar</ol><li>baz]</ol><p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<ol><ol><li>foo</li><ol><li>bar</li></ol><li>baz</li></ol></ol><p>extra</p>" not in array ["<blockquote><ol><li>foo</li><ol><li>bar</li></ol><li>baz</li></ol></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><ol><li>foo</li><ol><li>bar</li></ol><li>baz</li></ol></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<ol><li>[foo]<ol><li>bar</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li><li>bar</li></ol><li>baz</li></ol>" but got "<ol><ol><li>foo<ol><li>bar</li></ol>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>[bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><ol><li>bar</li></ol></ol><li>baz</li></ol>" but got "<ol><li>foo<ol><ol><li>bar</li></ol></ol>baz</li></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>bar</ol>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol></ol>" but got "<ol><ol><li>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>[foo<ol><li>bar]</ol>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li><ol><li>bar</li></ol></ol><li>baz</li></ol>" but got "<ol><ol><li>foo<ol><li>bar</li></ol>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<p>foo<!--bar-->{}<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p></blockquote><p>extra</p>" not in array ["<blockquote><p>foo<!--bar--></p></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo<!--bar--></p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<p>{}<!--foo-->bar<p>extra" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote><p>extra</p>" not in array ["<blockquote><p><!--foo-->bar</p></blockquote><p>extra</p>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p><!--foo-->bar</p></blockquote><p>extra</p>"]
[FAIL] [["indent",""]] "<blockquote><p>foo</blockquote> <p>[bar]" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote><p>foo</p></blockquote> <blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote>" not in array ["<blockquote><p>foo</p> <p>bar</p></blockquote>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p> <p>bar</p></blockquote>"]
[FAIL] [["indent",""]] "<p>[foo]</p> <blockquote><p>bar</blockquote>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p></blockquote> <blockquote><p>bar</p></blockquote>" not in array ["<blockquote><p>foo</p> <p>bar</p></blockquote>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p> <p>bar</p></blockquote>"]
[FAIL] [["indent",""]] "<blockquote><p>foo</blockquote> <p>[bar]</p> <blockquote><p>baz</blockquote>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<blockquote><p>foo</p></blockquote> <blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>bar</p></blockquote> <blockquote><p>baz</p></blockquote>" not in array ["<blockquote><p>foo</p> <p>bar</p> <p>baz</p></blockquote>", "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>foo</p> <p>bar</p> <p>baz</p></blockquote>"]
[FAIL] [["indent",""]] "<ol><li>foo</li><ol><li>bar</li></ol> <li>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol> </ol>"
[FAIL] [["indent",""]] "<ol><li>foo</li><ol><li>bar</li> </ol> <li>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li>  <li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol> </ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>bar</li> </ol></li><li>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol></ol>" but got "<ol><li>foo<ol><li>bar</li> </ol></li><ol><li>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>bar</li></ol></li> <li>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol></ol>" but got "<ol><li>foo<ol><li>bar</li></ol></li> <ol><li>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<ol><li>bar</li> </ol></li> <li>[baz]</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li>  <li>baz</li></ol></ol>" but got "<ol><li>foo<ol><li>bar</li> </ol></li> <ol><li>baz</li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<li>[bar]</li> <ol><li>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar</li><li>baz</li></ol> </ol>"
[FAIL] [["indent",""]] "<ol><li>foo<li>[bar]</li> <ol> <li>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li>  <li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol> </ol>"
[FAIL] [["indent",""]] "<ol><li>foo<li>[bar] <ol><li>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar </li><li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar <ol><li>baz</li></ol></li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<li>[bar]<ol> <li>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li> <li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar<ol> <li>baz</li></ol></li></ol></ol>"
[FAIL] [["indent",""]] "<ol><li>foo<li>[bar] <ol> <li>baz</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar </li> <li>baz</li></ol></ol>" but got "<ol><li>foo</li><ol><li>bar <ol> <li>baz</li></ol></li></ol></ol>"
[FAIL] [["indent",""]] "<ul><li>a<br>{<br>}</li><li>b</li></ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><ul><li>a<br><br></li></ul><li>b</li></ul>" but got "<ul><ul><li><br></li></ul><li>b</li></ul>"
Harness: the test ran to completion.
